# domeo-config-restore/Domeo_config-0952c5f19f4540c4fdb5582b721b7fe7517c6b37/app/api/price/doors/route.ts
import { NextRequest, NextResponse } from "next/server";
import { PrismaClient } from '@prisma/client';

const prisma = new PrismaClient();

// GET /api/price/doors - Получить базовую информацию о ценах
export async function GET(req: NextRequest) {
  try {
    const { searchParams } = new URL(req.url);
    const model = searchParams.get('model');
    
    if (!model) {
      return NextResponse.json({
        ok: true,
        message: "API для расчета цен дверей",
        usage: "Используйте POST запрос с данными selection для расчета цены",
        example: {
          method: "POST",
          body: {
            selection: {
              model: "Классика",
              hardware_kit: { id: "KIT_STD" },
              handle: { id: "HNDL_PRO" }
            }
          }
        }
      });
    }

    // Если передан model, возвращаем базовую информацию
    const product = await prisma.product.findFirst({
      where: { model },
      select: {
        id: true,
        sku: true,
        name: true,
        model: true,
        series: true,
        base_price: true
      }
    });

    if (!product) {
      return NextResponse.json({
        ok: false,
        error: "Продукт не найден"
      });
    }

    return NextResponse.json({
      ok: true,
      product: {
        id: product.id,
        sku: product.sku,
        name: product.name,
        model: product.model,
        series: product.series,
        base_price: product.base_price
      },
      message: "Для полного расчета цены используйте POST запрос"
    });
  } catch (error) {
    console.error('Error in GET /api/price/doors:', error);
    return NextResponse.json(
      { error: "Ошибка получения информации о ценах" },
      { status: 500 }
    );
  }
}

export async function POST(req: NextRequest) {
  try {
    const body = await req.json();
    console.log('📦 Raw request body:', JSON.stringify(body, null, 2));
    console.log('🔍 Body type:', typeof body);
    console.log('🔍 Body.selection:', body?.selection);
    
    // Данные могут приходить напрямую в body или в поле selection
    const selection = body?.selection || body;
    console.log('🎯 Extracted selection:', JSON.stringify(selection, null, 2));
    
    console.log('💰 Price calculation request:', {
      style: selection?.style,
      model: selection?.model,
      finish: selection?.finish,
      color: selection?.color,
      width: selection?.width,
      height: selection?.height,
      hardware_kit: selection?.hardware_kit,
      handle: selection?.handle
    });
    

    if (!selection) {
      console.error('❌ Selection is undefined or null');
      return NextResponse.json(
        { error: "Данные для расчета не предоставлены" },
        { status: 400 }
      );
    }

    // Ищем продукт в базе данных по всем параметрам
    const products = await prisma.product.findMany({
      where: {
        catalog_category: {
          name: "Межкомнатные двери"
        }
      },
      select: {
        id: true,
        sku: true,
        name: true,
        model: true,
        series: true,
        base_price: true,
        properties_data: true
      },
      orderBy: {
        id: 'asc'
      }
    });

    // Фильтруем товары по выбранным параметрам
    const matchingProducts = products.filter(p => {
      const properties = p.properties_data ? 
        (typeof p.properties_data === 'string' ? JSON.parse(p.properties_data) : p.properties_data) : {};

      // Сначала ищем по стилю и модели
      const styleMatch = !selection.style || properties['Domeo_Стиль Web'] === selection.style;
      
      // Ищем по полному названию модели (как в калькуляторе)
      const modelMatch = !selection.model || properties['Domeo_Название модели для Web'] === selection.model;
      
      if (!styleMatch || !modelMatch) {
        return false;
      }

      // Затем по остальным параметрам
      const finishMatch = !selection.finish || properties['Тип покрытия'] === selection.finish;
      const colorMatch = !selection.color || properties['Domeo_Цвет'] === selection.color;
      const typeMatch = !selection.type || properties['Тип конструкции'] === selection.type;
      const widthMatch = !selection.width || properties['Ширина/мм'] == selection.width;
      const heightMatch = !selection.height || properties['Высота/мм'] == selection.height;
      
      console.log('🔍 Product filter:', {
        productId: p.id,
        styleMatch,
        modelMatch,
        finishMatch,
        colorMatch,
        typeMatch,
        widthMatch,
        heightMatch,
        requestedModel: selection.model,
        actualModelName: properties['Domeo_Название модели для Web'],
        actualSupplierSku: properties['Артикул поставщика'],
        requestedFinish: selection.finish,
        actualFinish: properties['Тип покрытия'],
        requestedColor: selection.color,
        actualColor: properties['Domeo_Цвет'],
        requestedWidth: selection.width,
        actualWidth: properties['Ширина/мм'],
        requestedHeight: selection.height,
        actualHeight: properties['Высота/мм']
      });
      
      return finishMatch && colorMatch && typeMatch && widthMatch && heightMatch;
    });

    console.log(`📦 Найдено ${matchingProducts.length} подходящих товаров`);

    // Если товары не найдены, возвращаем ошибку
    if (matchingProducts.length === 0) {
      console.log('❌ Точный товар не найден для параметров:', {
        style: selection.style,
        model: selection.model,
        finish: selection.finish,
        color: selection.color,
        width: selection.width,
        height: selection.height
      });
      
      return NextResponse.json(
        { error: "Товар с указанными параметрами не найден" },
        { status: 404 }
      );
    }

    // Выбираем товар с МАКСИМАЛЬНОЙ ценой (временно)
    const product = matchingProducts.reduce((maxProduct, currentProduct) => {
      const maxProps = maxProduct.properties_data ? 
        (typeof maxProduct.properties_data === 'string' ? JSON.parse(maxProduct.properties_data) : maxProduct.properties_data) : {};
      const currentProps = currentProduct.properties_data ? 
        (typeof currentProduct.properties_data === 'string' ? JSON.parse(currentProduct.properties_data) : currentProduct.properties_data) : {};
      
      const maxPrice = parseFloat(maxProps['Цена РРЦ']) || maxProduct.base_price || 0;
      const currentPrice = parseFloat(currentProps['Цена РРЦ']) || currentProduct.base_price || 0;
      
      console.log(`💰 Сравниваем цены: ${maxProduct.sku} (${maxPrice} руб.) vs ${currentProduct.sku} (${currentPrice} руб.)`);
      
      return currentPrice > maxPrice ? currentProduct : maxProduct;
    }, matchingProducts[0]);
    
    const finalProduct = product;

    // Парсим свойства продукта
    const properties = finalProduct.properties_data ? 
      (typeof finalProduct.properties_data === 'string' ? JSON.parse(finalProduct.properties_data) : finalProduct.properties_data) : {};

    // Рассчитываем цену из цены РРЦ товара
    const rrcPrice = parseFloat(properties['Цена РРЦ']) || 0;
    const basePrice = finalProduct.base_price || 0;
    let doorPrice = rrcPrice || basePrice;
    
    console.log('💰 Price calculation:', {
      productId: finalProduct.id,
      rrcPrice,
      basePrice,
      finalDoorPrice: doorPrice,
      rrcPriceExists: !!properties['Цена РРЦ']
    });
    
    let total = doorPrice;
    const breakdown = [
      { label: "Дверь", amount: doorPrice }
    ];

    // Добавляем комплект фурнитуры если выбран
    if (selection.hardware_kit?.id) {
      console.log('🔧 Hardware kit selected:', selection.hardware_kit.id);
      
      // Получаем комплекты фурнитуры из базы данных
      const hardwareKits = await prisma.product.findMany({
        where: {
          catalog_category: {
            name: "Комплекты фурнитуры"
          }
        },
        select: {
          id: true,
          name: true,
          properties_data: true
        }
      });

      console.log('🔧 Available hardware kits:', hardwareKits.length);
      const kit = hardwareKits.find(k => k.id === selection.hardware_kit.id);
      console.log('🔧 Found kit:', kit ? 'Yes' : 'No');
      if (kit) {
        const kitProps = kit.properties_data ? 
          (typeof kit.properties_data === 'string' ? JSON.parse(kit.properties_data) : kit.properties_data) : {};
        
        const kitPrice = parseFloat(kitProps['Группа_цена']) || 0;
        console.log('🔧 Kit price:', kitPrice);
        total += kitPrice;
        breakdown.push({ 
          label: `Комплект: ${kitProps['Наименование для Web'] || kit.name}`, 
          amount: kitPrice 
        });
      }
    }

    // Добавляем ручку если выбрана
    if (selection.handle?.id) {
      console.log('🔧 Handle selected:', selection.handle.id);
      
      // Получаем ручки из базы данных
      const handles = await prisma.product.findMany({
        where: {
          catalog_category: {
            name: "Ручки"
          }
        },
        select: {
          id: true,
          name: true,
          properties_data: true
        }
      });

      console.log('🔧 Available handles:', handles.length);
      const handle = handles.find(h => h.id === selection.handle.id);
      console.log('🔧 Found handle:', handle ? 'Yes' : 'No');
      if (handle) {
        const handleProps = handle.properties_data ? 
          (typeof handle.properties_data === 'string' ? JSON.parse(handle.properties_data) : handle.properties_data) : {};
        
        const handlePrice = parseFloat(handleProps['Domeo_цена группы Web']) || 0;
        console.log('🔧 Handle price:', handlePrice);
        total += handlePrice;
        breakdown.push({ 
          label: `Ручка: ${handleProps['Domeo_наименование ручки_1С'] || handle.name}`, 
          amount: handlePrice 
        });
      }
    }

    const result = {
      ok: true,
      currency: "RUB",
      base: doorPrice,
      breakdown,
      total: Math.round(total),
      sku: finalProduct.sku
    };
    
    
    return new NextResponse(JSON.stringify(result), {
      status: 200,
      headers: {
        'Content-Type': 'application/json; charset=utf-8',
      },
    });
  } catch (error) {
    console.error('Error calculating door price:', error);
    return NextResponse.json(
      { error: "Ошибка расчета цены" },
      { status: 500 }
    );
  }
}